refactor(builder): extrahiere diagnosis.rs aus single/mod.rs
- build_finding_summary, build_thematic_clusters, dimension_cluster_label,
  build_diagnosis_block → single/diagnosis.rs
- mod.rs schrumpft auf ~1.694 Zeilen
